refactor: 타 컴포넌트에서 사용되는 인터페이스를 따로 빼서 관리
